@@ -25,20 +25,19 @@
         "files": [ "src/exceptions.php" ]
     },
     "require": {
-        "php": ">=5.5.0",
+        "php": ">=8.1",
         "ext-gettext": "*",
         "ext-hash": "*",
         "ext-json": "*",
-        "guzzlehttp/guzzle": "~6.3"
+        "guzzlehttp/guzzle": "~7.4"
     },
     "require-dev": {
         "pdepend/pdepend": "^2.2",
         "phpmd/phpmd": "^2.4",
-        "phploc/phploc": "^2.1",
+        "phploc/phploc": "^7.0",
         "sebastian/phpcpd": "^2.0",
-        "phpdocumentor/phpdocumentor": "^2.9",
-        "mamuz/php-dependency-analysis": "^0.6",
-        "phpunit/phpunit": "^5.7",
-        "friendsofphp/php-cs-fixer": "^1.0"
+        "phpdocumentor/phpdocumentor": "3.1",
+        "phpunit/phpunit": "^9.5",
+        "friendsofphp/php-cs-fixer": "3.4"
     }
 }
